can now work as a node module or a web project module
